requirements.txt
gitpython>=3.1.0
requests>=2.25.0
watchdog>=2.1.0